Revert "PIM-5853: Product identifier filter"
