+++ src/Pim/Bundle/EnrichBundle/Resources/config/form_extensions/filters.yml
@@ -28,8 +28,5 @@ extensions:
                 - ">="
                 - SINCE LAST EXPORT
 
-    pim-filter-product-identifier:
-        module: pim/filter/product/identifier
-
     pim-filter-simpleselect:
         module: pim/filter/simpleselect
