+++ src/Pim/Bundle/EnrichBundle/Resources/config/form_extensions/product_export_show.yml
@@ -46,9 +46,6 @@ extensions:
                 -
                     field: updated
                     view: pim-filter-product-updated
-                -
-                    field: id
-                    view: pim-filter-product-identifier
 
     pim-product-export-show-content-data-help:
         module: pim/export/product/edit/content/data/help
