+++ src/Pim/Bundle/EnrichBundle/Resources/config/requirejs.yml
@@ -227,14 +227,13 @@ config:
         pim/export/product/edit/content/data/help:                 pimenrich/js/export/product/edit/content/data/help
 
         # Filters
-        pim/filter/filter:               pimenrich/js/filter/filter
-        pim/filter/text:                 pimenrich/js/filter/text
-        pim/filter/simpleselect:         pimenrich/js/filter/simpleselect
-        pim/filter/product/family:       pimenrich/js/filter/product/family
-        pim/filter/product/completeness: pimenrich/js/filter/product/completeness
-        pim/filter/product/enabled:      pimenrich/js/filter/product/enabled
-        pim/filter/product/updated:      pimenrich/js/filter/product/updated
-        pim/filter/product/identifier:   pimenrich/js/filter/product/identifier
+        pim/filter/filter:                pimenrich/js/filter/filter
+        pim/filter/text:                  pimenrich/js/filter/text
+        pim/filter/simpleselect:          pimenrich/js/filter/simpleselect
+        pim/filter/product/family:        pimenrich/js/filter/product/family
+        pim/filter/product/completeness:  pimenrich/js/filter/product/completeness
+        pim/filter/product/enabled:       pimenrich/js/filter/product/enabled
+        pim/filter/product/updated:       pimenrich/js/filter/product/updated
 
         # Product edit form
         pim/field-manager:                  pimenrich/js/product/field-manager
@@ -385,4 +384,3 @@ config:
         pim/template/filter/product/enabled:      pimenrich/templates/filter/product/enabled.html
         pim/template/filter/product/updated:      pimenrich/templates/filter/product/updated.html
         pim/template/filter/product/completeness: pimenrich/templates/filter/product/completeness.html
-        pim/template/filter/product/identifier:   pimenrich/templates/filter/product/identifier.html
