--- src/Pim/Bundle/EnrichBundle/Resources/public/templates/filter/product/identifier.html
@@ -1,7 +1,0 @@
-<textarea
-    class="identifier input-large"
-    name="filter-value"
-    <%- isEditable ? '' : 'disabled' %>
->
-<%- value %>
-</textarea>
